API Swap SSTemplateParserProxy for SSViewerProxy
SSViewer is now injectable, which means that we can list used templates outside of the cache
